pyproject.toml
[project]
name = "astera-demogorgn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "icechunk>=1.1.21",
    "matplotlib>=3.10.8",
    "netcdf4>=1.7.4",
    "scipy>=1.17.1",
    "xarray>=2026.2.0",
    "zarr>=3.1.6",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]

[project.optional-dependencies]
ci = [
    "boto3>=1.34",
    "pyyaml>=6.0",
]
